+++ app/models/concerns/replyable.rb
@@ -7,6 +7,7 @@ module Replyable
   included do
     # Relationships
     has_many :replies, -> { order :id }
+    has_many :countable_replies, -> { current.shadow_banned(nil) }, class_name: "Reply", source: :reply
     has_many :reply_users
   end
 
@@ -14,6 +15,5 @@ module Replyable
     ((replies.where(reply_id: nil).count - 1) / Reply::REPLIES_PER_PAGE) + 1
   end
 
-  def get_or_create_subscription(current_user)
-  end
+  def get_or_create_subscription(current_user); end
 end