+++ app/models/concerns/searchable.rb
@@ -1,20 +1,25 @@
 # frozen_string_literal: true
 
-# Allows models to be search based on certain attributes
+# Allows models to be searched.
 module Searchable
   extend ActiveSupport::Concern
 
   included do
     # Search Scope
-    def self.search(arg)
-      term = arg.to_s.downcase.gsub(/^| |$/, '%')
+    def self.search(arg, match_start: true)
+      term = \
+        if match_start
+          arg.to_s.downcase.gsub(/ |$/, "%")
+        else
+          arg.to_s.downcase.gsub(/^| |$/, "%")
+        end
       terms = [term] * search_queries.count
-      where search_queries.join(' or '), *terms
+      where search_queries.join(" or "), *terms
     end
 
     def self.search_queries
       searchable_attributes.collect do |searchable_attribute|
-        "LOWER(#{table_name}.#{searchable_attribute}) LIKE ?"
+        "#{table_name}.#{searchable_attribute} ILIKE ?"
       end
     end