+++ config/routes.rb
@@ -21,6 +21,13 @@ Rails.application.routes.draw do
   get "admin" => "admin#dashboard", as: :admin
 
   namespace :admin do
+    get :spam_inbox, path: "spam-inbox"
+    get :spam_report, path: "spam-report(/:year)"
+    get :profile_review, path: "profile-review"
+    post :submit_profile_review, path: "profile-review"
+    post :unspamban, path: "unspamban/:id"
+    post :destroy_spammer, path: "empty-spam/:id"
+    post :empty_spam, path: "empty-spam"
     get :roles
     get :stats
     get :sync